requirements.txt
einops==0.8.1
jaxtyping==0.3.2
transformers==4.54.1
diffusers==0.34.0
tokenizers==0.21.4
peft==0.16.0
sentencepiece==0.2.0
bitsandbytes==0.46.1
protobuf==6.31.1
webdataset==1.0.2
torchcodec==0.5
fire
wandb